+++ lang/src/accounts/mod.rs
@@ -24,3 +24,5 @@ pub mod state;
 pub mod system_account;
 pub mod sysvar;
 pub mod unchecked_account;
+
+pub use anchor_common::header;