+++ lang/src/bpf_upgradeable_state.rs
@@ -37,12 +37,7 @@ impl AccountDeserialize for ProgramData {
     }
 }
 
-impl AccountSerialize for ProgramData {
-    fn try_serialize<W: std::io::Write>(&self, _writer: &mut W) -> Result<()> {
-        // no-op
-        Ok(())
-    }
-}
+impl AccountSerialize for ProgramData {}
 
 impl Owner for ProgramData {
     fn owner() -> solana_program::pubkey::Pubkey {
@@ -56,12 +51,7 @@ impl Owner for UpgradeableLoaderState {
     }
 }
 
-impl AccountSerialize for UpgradeableLoaderState {
-    fn try_serialize<W: std::io::Write>(&self, _writer: &mut W) -> Result<()> {
-        // no-op
-        Ok(())
-    }
-}
+impl AccountSerialize for UpgradeableLoaderState {}
 
 impl AccountDeserialize for UpgradeableLoaderState {
     fn try_deserialize(buf: &mut &[u8]) -> Result<Self> {