+++ lang/syn/src/codegen/program/handlers.rs
@@ -89,7 +89,7 @@ pub fn generate(program: &Program) -> proc_macro2::TokenStream {
                 let seed = anchor_lang::idl::IdlAccount::seed();
                 let owner = accounts.program.key;
                 let to = Pubkey::create_with_seed(&base, seed, owner).unwrap();
-                // Space: account discriminator || authority pubkey || vec len || vec data
+                // Space: account header || authority pubkey || vec len || vec data
                 let space = 8 + 32 + 4 + data_len as usize;
                 let rent = Rent::get()?;
                 let lamports = rent.minimum_balance(space);
@@ -128,9 +128,7 @@ pub fn generate(program: &Program) -> proc_macro2::TokenStream {
 
                 // Store the new account data.
                 let mut data = accounts.to.try_borrow_mut_data()?;
-                let dst: &mut [u8] = &mut data;
-                let mut cursor = std::io::Cursor::new(dst);
-                idl_account.try_serialize(&mut cursor)?;
+                idl_account.try_serialize(&mut data)?;
 
                 Ok(())
             }
@@ -201,6 +199,16 @@ pub fn generate(program: &Program) -> proc_macro2::TokenStream {
                 let ix_name: proc_macro2::TokenStream =
                     generate_ctor_variant_name().parse().unwrap();
                 let ix_name_log = format!("Instruction: {}", ix_name);
+                let header_write = quote! {
+                    {
+                        use anchor_lang::Discriminator;
+                        let mut __data = ctor_accounts.to.try_borrow_mut_data()?;
+                        anchor_lang::accounts::header::init(
+                            &mut __data,
+                            &#name::discriminator(),
+                        );
+                    }
+                };
                 if state.is_zero_copy {
                     quote! {
                         // One time state account initializer. Will faill on subsequent
@@ -254,14 +262,17 @@ pub fn generate(program: &Program) -> proc_macro2::TokenStream {
                                 &[seeds],
                             )?;
 
+                            // Initialize the header.
+                            #header_write
+
                             // Zero copy deserialize.
-                            let loader: anchor_lang::accounts::loader::Loader<#mod_name::#name> = anchor_lang::accounts::loader::Loader::try_from_unchecked(program_id, &ctor_accounts.to)?;
+                            let loader: anchor_lang::accounts::loader::Loader<#mod_name::#name> = anchor_lang::accounts::loader::Loader::try_from(program_id, &ctor_accounts.to)?;
 
                             // Invoke the ctor in a new lexical scope so that
                             // the zero-copy RefMut gets dropped. Required
                             // so that we can subsequently run the exit routine.
                             {
-                                let mut instance = loader.load_init()?;
+                                let mut instance = loader.load_mut()?;
                                 instance.new(
                                     anchor_lang::context::Context::new(
                                         program_id,
@@ -344,12 +355,13 @@ pub fn generate(program: &Program) -> proc_macro2::TokenStream {
                                 &[seeds],
                             )?;
 
+                            // Initialize the account header.
+                            #header_write
+
                             // Serialize the state and save it to storage.
                             ctor_user_def_accounts.exit(program_id)?;
                             let mut data = ctor_accounts.to.try_borrow_mut_data()?;
-                            let dst: &mut [u8] = &mut data;
-                            let mut cursor = std::io::Cursor::new(dst);
-                            instance.try_serialize(&mut cursor)?;
+                            instance.try_serialize(&mut data)?;
 
                             Ok(())
                         }
@@ -498,9 +510,7 @@ pub fn generate(program: &Program) -> proc_macro2::TokenStream {
                                     accounts.exit(program_id)?;
                                     let acc_info = state.to_account_info();
                                     let mut data = acc_info.try_borrow_mut_data()?;
-                                    let dst: &mut [u8] = &mut data;
-                                    let mut cursor = std::io::Cursor::new(dst);
-                                    state.try_serialize(&mut cursor)?;
+                                    state.try_serialize(&mut data)?;
 
                                     Ok(())
                                 }
@@ -626,9 +636,7 @@ pub fn generate(program: &Program) -> proc_macro2::TokenStream {
                                             accounts.exit(program_id)?;
                                             let acc_info = state.to_account_info();
                                             let mut data = acc_info.try_borrow_mut_data()?;
-                                            let dst: &mut [u8] = &mut data;
-                                            let mut cursor = std::io::Cursor::new(dst);
-                                            state.try_serialize(&mut cursor)?;
+                                            state.try_serialize(&mut data)?;
 
                                             Ok(())
                                         }