+++ ts/src/coder/borsh/index.ts
@@ -6,9 +6,9 @@ import { BorshStateCoder } from "./state.js";
 import { Coder } from "../index.js";
 
 export { BorshInstructionCoder } from "./instruction.js";
-export { BorshAccountsCoder, ACCOUNT_DISCRIMINATOR_SIZE } from "./accounts.js";
-export { BorshEventCoder, eventDiscriminator } from "./event.js";
-export { BorshStateCoder, stateDiscriminator } from "./state.js";
+export { BorshAccountsCoder, BorshAccountHeader } from "./accounts.js";
+export { BorshEventCoder } from "./event.js";
+export { BorshStateCoder } from "./state.js";
 
 /**
  * BorshCoder is the default Coder for Anchor programs implementing the